requirements.txt
# Core ML and Vector Operations
transformers>=4.0.0
sentence-transformers>=2.0.0
qdrant-client>=1.0.0
torch>=1.9.0

# GitHub Integration
requests>=2.25.0
PyGithub>=1.55.0

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Testing Framework
pytest>=6.0.0
hypothesis>=6.0.0
pytest-cov>=2.12.0
pytest-asyncio>=0.15.0

# Development Tools
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Configuration and Logging
pyyaml>=5.4.0
python-dotenv>=0.19.0
structlog>=21.1.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0

# Utilities
python-multipart>=0.0.5
aiofiles>=0.7.0